+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Imported user guide from http://agilesoc.com/open-source-projects/svunit/svunit-user-guide/ into
   repo
 - Support for DSim
+- `--filter` option for `runSVUnit` to control which tests should run
 
 ### Fixed
 - Fixed FAIL_IF and FAIL_UNLESS macros to properly implement SystemVerilog